moves cache entry to a separate file
